Added scripts to calculate derived CMIP6 variables
Modified the split file derived variable code used with
the split BCCAQv2 output files so that it can work with
the complete CMIP6 GCM files. These take the assembled
GCM files and compute derived variables at the same
resolution.
Also added code to run TX90p (and other 3 vars) that
require the Climdex Object be computed. This was not
part of the split file derivation.